requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
pydot==1.4.2
pyparsing==3.0.9
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
